refactor(cli): extract doctor command to cli_doctor.py (ADR-059 V9)
